+++ test/jasmine/tests/gl_plot_interact_test.js
@@ -211,70 +211,6 @@ describe('Test gl plot interactions', function() {
 
     });
 
-    describe('gl3d plots', function() {
-
-        // Expected shape of projection-related data
-        var cameraStructure = {
-            up: {x: jasmine.any(Number), y: jasmine.any(Number), z: jasmine.any(Number)},
-            center: {x: jasmine.any(Number), y: jasmine.any(Number), z: jasmine.any(Number)},
-            eye: {x: jasmine.any(Number), y: jasmine.any(Number), z: jasmine.any(Number)}
-        };
-
-        function makePlot(mock) {
-            return Plotly.plot(createGraphDiv(), mock.data, mock.layout);
-        }
-
-        function addEventCallback(graphDiv) {
-            var relayoutCallback = jasmine.createSpy('relayoutCallback');
-            graphDiv.on('plotly_relayout', relayoutCallback);
-            return {graphDiv: graphDiv, relayoutCallback: relayoutCallback};
-        }
-
-        function verifyInteractionEffects(tuple) {
-
-            // One 'drag': simulating fairly thoroughly as the mouseup event is also needed here
-            mouseEvent('mousemove', 400, 200);
-            mouseEvent('mousedown', 400, 200);
-            mouseEvent('mousemove', 320, 320, {buttons: 1});
-            mouseEvent('mouseup', 320, 320);
-
-            // Check event emission count
-            expect(tuple.relayoutCallback).toHaveBeenCalledTimes(1);
-
-            // Check structure of event callback value contents
-            expect(tuple.relayoutCallback).toHaveBeenCalledWith(jasmine.objectContaining({scene: cameraStructure}));
-
-            // Check camera contents on the DIV layout
-            var divCamera = tuple.graphDiv.layout.scene.camera;
-
-            expect(divCamera).toEqual(cameraStructure);
-
-            return tuple.graphDiv;
-        }
-
-        function markForTeardown(graphDiv) {
-            // TODO consider changing this test file such that the teardown needs no communication via a global variable
-            gd = graphDiv;
-        }
-
-        function testEvents(plot, done) {
-            plot.then(function(graphDiv) {
-                var tuple = addEventCallback(graphDiv);
-                verifyInteractionEffects(tuple);
-                markForTeardown(graphDiv);
-                done();
-            });
-        }
-
-        it('should respond to drag interactions with mock of unset camera', function(done) {
-            testEvents(makePlot(require('@mocks/gl3d_scatter3d-connectgaps.json')), done);
-        });
-
-        it('should respond to drag interactions with mock of partially set camera', function(done) {
-            testEvents(makePlot(require('@mocks/gl3d_errorbars_zx.json')), done);
-        });
-    });
-
     describe('gl2d plots', function() {
         var mock = require('@mocks/gl2d_10.json'),
             modeBar, relayoutCallback;
